RFC: new option: --dsym
If a Formula specifies "option :dsym", and the user specifies '--dsym' on the
command line, then debug symbols will be built and included in the Keg.

This relies on the Apple "lazy dwarf scheme", which leaves the debug info in the
object files, with only links to the object files present in the final
exe/dylib.  The 'dsymutil' tool reads the final exe, finds the objects, and
links all the debug info into a .dSYM bundle.

http://wiki.dwarfstd.org/index.php?title=Apple's_%22Lazy%22_DWARF_Scheme

This also adds a new formula method 'install_dsym', which is responsible for
finding all the MACHO files in the keg and running dsymutil on them.  It
provides a default implementation, which may be overridden if need be.

Signed-off-by: Lawrence D'Anna <larry@elder-gods.org>